feat: inject bootstrap jobs for unonboarded repos on PR
