+++ cmd/in-repo-config-plugin/config_test.go
