+++ cmd/in-repo-config-plugin/main.go
@@ -5,10 +5,8 @@ import (
 	"flag"
 	"fmt"
 	"os"
-	"runtime"
 	"time"
 
-	"github.com/prometheus/client_golang/prometheus"
 	"github.com/sirupsen/logrus"
 
 	pjapi "sigs.k8s.io/prow/pkg/apis/prowjobs/v1"
@@ -19,95 +17,14 @@ import (
 	"sigs.k8s.io/prow/pkg/interrupts"
 	"sigs.k8s.io/prow/pkg/logrusutil"
 	"sigs.k8s.io/prow/pkg/pjutil"
+
+	"github.com/openshift/ci-tools/pkg/dispatcher"
 	k8sruntime "k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/rest"
 	ctrlruntimeclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 const pluginName = "in-repo-config"
-
-var (
-	concurrentHandlersInFlight = prometheus.NewGauge(prometheus.GaugeOpts{
-		Name: "in_repo_config_concurrent_handlers",
-		Help: "Number of webhook handlers currently executing",
-	})
-	droppedHandlerRequests = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Name: "in_repo_config_dropped_requests_total",
-		Help: "Number of webhook requests dropped",
-	}, []string{"reason"})
-)
-
-func init() {
-	prometheus.MustRegister(concurrentHandlersInFlight)
-	prometheus.MustRegister(droppedHandlerRequests)
-}
-
-type handlerDispatcher struct {
-	executionSlots   chan struct{}
-	queueSlots       chan struct{}
-	queueTimeout     time.Duration
-	executionTimeout time.Duration
-}
-
-func newHandlerDispatcher(maxConcurrent, maxQueued int, queueTimeout, executionTimeout time.Duration) *handlerDispatcher {
-	return &handlerDispatcher{
-		executionSlots:   make(chan struct{}, maxConcurrent),
-		queueSlots:       make(chan struct{}, maxQueued),
-		queueTimeout:     queueTimeout,
-		executionTimeout: executionTimeout,
-	}
-}
-
-func (d *handlerDispatcher) dispatch(logger *logrus.Entry, handler func()) {
-	select {
-	case d.executionSlots <- struct{}{}:
-		d.run(logger, handler)
-		return
-	default:
-	}
-
-	select {
-	case d.queueSlots <- struct{}{}:
-	default:
-		droppedHandlerRequests.WithLabelValues("queue_full").Inc()
-		logger.Warn("dropping webhook request: handler queue is full")
-		return
-	}
-
-	select {
-	case d.executionSlots <- struct{}{}:
-		<-d.queueSlots
-		d.run(logger, handler)
-	case <-time.After(d.queueTimeout):
-		<-d.queueSlots
-		droppedHandlerRequests.WithLabelValues("queue_timeout").Inc()
-		logger.Warn("dropping webhook request: waited too long in queue")
-	}
-}
-
-func (d *handlerDispatcher) run(logger *logrus.Entry, handler func()) {
-	done := make(chan struct{})
-	go func() {
-		concurrentHandlersInFlight.Inc()
-		defer concurrentHandlersInFlight.Dec()
-		defer func() { <-d.executionSlots }()
-		defer close(done)
-		defer func() {
-			if r := recover(); r != nil {
-				stack := make([]byte, 8192)
-				stack = stack[:runtime.Stack(stack, false)]
-				logger.Errorf("webhook handler panicked: %v\n%s", r, stack)
-			}
-		}()
-		handler()
-	}()
-
-	select {
-	case <-done:
-	case <-time.After(d.executionTimeout):
-		logger.Warn("webhook handler exceeded execution timeout, still running in background")
-	}
-}
 
 type options struct {
 	logLevel                 string
@@ -119,6 +36,7 @@ type options struct {
 	checkconfigImage         string
 	prowConfigPath           string
 	namespace                string
+	dispatcherURL            string
 	maxConcurrentHandlers    int
 	maxQueuedHandlers        int
 	queueTimeoutMinutes      int
@@ -136,6 +54,7 @@ func gatherOptions() options {
 	fs.StringVar(&o.prowgenImage, "prowgen-image", "", "Container image for ci-operator-prowgen used in bootstrap jobs.")
 	fs.StringVar(&o.checkconfigImage, "checkconfig-image", "", "Container image for ci-operator-checkconfig used in the config-check presubmit.")
 	fs.StringVar(&o.prowConfigPath, "prow-config-path", "/etc/config/config.yaml", "Path to the Prow config file for ProwJob defaulting.")
+	fs.StringVar(&o.dispatcherURL, "dispatcher-url", "", "URL of the prow-job-dispatcher for cluster assignment. When set, all triggered jobs get their cluster from this service.")
 	fs.StringVar(&o.namespace, "namespace", "ci", "Namespace where ProwJobs will be created.")
 	fs.IntVar(&o.maxConcurrentHandlers, "max-concurrent-handlers", 3, "Maximum number of webhook handlers running concurrently.")
 	fs.IntVar(&o.maxQueuedHandlers, "max-queued-handlers", 20, "Maximum number of webhook handlers waiting in queue.")
@@ -213,6 +132,12 @@ func main() {
 		logger.WithError(err).Fatal("Error creating ProwJob client.")
 	}
 
+	var dispatcherClient dispatcher.Client
+	if o.dispatcherURL != "" {
+		dispatcherClient = dispatcher.NewClient(o.dispatcherURL)
+		logger.WithField("url", o.dispatcherURL).Info("Using dispatcher for cluster assignment")
+	}
+
 	serv := &server{
 		ghc:              githubClient,
 		pjclient:         pjclient,
@@ -221,18 +146,19 @@ func main() {
 		jobConfigDir:     o.jobConfigDir,
 		prowgenImage:     o.prowgenImage,
 		checkconfigImage: o.checkconfigImage,
+		dispatcher:       dispatcherClient,
 	}
 
 	queueTimeout := time.Duration(o.queueTimeoutMinutes) * time.Minute
 	executionTimeout := time.Duration(o.handlerTimeoutMinutes) * time.Minute
-	dispatcher := newHandlerDispatcher(o.maxConcurrentHandlers, o.maxQueuedHandlers, queueTimeout, executionTimeout)
+	throttler := newWebhookThrottler(o.maxConcurrentHandlers, o.maxQueuedHandlers, queueTimeout, executionTimeout)
 
 	eventServer := githubeventserver.New(o.githubEventServerOptions, getWebhookHMAC, logger)
 	eventServer.RegisterHandlePullRequestEvent(func(l *logrus.Entry, event github.PullRequestEvent) {
-		dispatcher.dispatch(l, func() { serv.handlePullRequest(l, event) })
+		throttler.handle(l, func() { serv.handlePullRequest(l, event) })
 	})
 	eventServer.RegisterPushEventHandler(func(l *logrus.Entry, event github.PushEvent) {
-		dispatcher.dispatch(l, func() { serv.handlePush(l, event) })
+		throttler.handle(l, func() { serv.handlePush(l, event) })
 	})
 	eventServer.RegisterHelpProvider(helpProvider, logger)
