+++ cmd/in-repo-config-plugin/templates/config-check-presubmit.yaml.tmpl
@@ -11,7 +11,7 @@ decorate: true
 extra_refs:
   - org: openshift
     repo: release
-    base_ref: master
+    base_ref: main
 spec:
   containers:
     - name: checkconfig
