+++ images/staging-dispatcher/Dockerfile
@@ -1,0 +1,3 @@
+FROM registry.access.redhat.com/ubi9/ubi-minimal:latest
+ADD staging-dispatcher /usr/bin/staging-dispatcher
+ENTRYPOINT ["/usr/bin/staging-dispatcher"]
